+++ packages/form/src/components/ModalForm/demos/modal-form-submitter.tsx
@@ -104,6 +104,21 @@ export default () => {
 
         <ProFormText width="md" name="company" label="我方公司名称" placeholder="请输入名称" />
       </ModalForm>
+      <ModalForm
+        title="隐藏footer"
+        trigger={<Button type="primary">隐藏footer</Button>}
+        submitter={false}
+      >
+        <ProFormText
+          width="md"
+          name="name"
+          label="签约客户名称"
+          tooltip="最长为 24 位"
+          placeholder="请输入名称"
+        />
+
+        <ProFormText width="md" name="company" label="我方公司名称" placeholder="请输入名称" />
+      </ModalForm>
     </Space>
   );
 };