+++ packages/list/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.9.0...@ant-design/pro-list@1.9.1) (2021-06-09)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 # [1.9.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.8.2...@ant-design/pro-list@1.9.0) (2021-06-08)
 
 ### Features